Ok, I have added Entcover stuff. I couldn't actually put ent_make_struct.f into a separate directory because it uses stuff from Ent and is itself used by ent_mod. So I had to put it into Ent/ . I have created a directory Entcover_standalone and moved Entcover driver there. (I left Entcover/ untouched just for a reference).
The reading routine had to be broken into two parts. The main part
(which reads entcell structure) was left in ent_make_struct.f
while the part which opens the file and reads i,j was moved
to ent_struct.f . That's because Ent below ent_mod doesn't
know anything about i,j (it operates on single cells only).
The only change in format this causes is that now one has
to put "$" at the end of each entcell. So at the end of file
the will be two $ (one for cell and one for file).
One has to add and extra option NEED_Entcover=YES to Ent options
(to tell Ent to compile ent_make_struct.f).

The model currently compiles for r_ent_cover.mk but probably
requires debugging.

I may have missed some files in this commit. Will commit the
rest in a moment.