phenology.f : fixed C conservation problem in litter_turnover_cohort. Commented out "is_annual" part since it doesn't conserve carbon and is causing other problems. Also added some C conservation diags.
GHY.f: returned to the "old" way of calling vegetation update
from inside the time loop. (just to make sure that the difference
with the original code is minimal).

ent.f: C diags

entcells.f: removed "pointer" attribute of the argument in
entcell_carbon (why was it there ???)

This code seems to run ok for the first ~ 30 years (if one
considers it to be ok to loose all LAI in Siberia), then the
results (for C pools) become completely unrealistic and finally
turn into NaN's.