Introducing a more general treatment of orbits and calendars.
These changes are primarily for inctroducing a more nuanced calendar
for non-Earth planets in which the length of the seasons is properly
reflected in the length of the months.  Calendars attempt to preserve
the month angles of the Julian calendar in the context of other orbits.

Certain changes were necessary to isolate some Earth-specific logic in
the procedure ORBPAR() which has now been moved to the ./model/shared
directory.

A number of clasess have been introduced to facilitate all of this.

Most importantly there are AbstractCalendard and AbstractOrbit classes
whose intent should be clear from their names.  Each concrete Orbit
subclass is expected to implement the makeCalendar() method.  In this
way calendars are guaranteed to be consistent with the orbit from
which they are derived.

To support the existing Earth logic, there remains the JulianCalendar
class.  There is also the ParameterizedEarthOrbit which updates
orbital parameters annualy via ORBPAR().  Finally for static Earth,
there is a simpler subclass: Earth365DayOrbit.

For other planets there is a PlanetaryOrbit class and a
PlanetaryCalendar class.  The primary limitation is that each month
must have at least 1 day.  I.e these implementations are no
appropriate for tidally locked planets or even planets with ~12 or
fewer days per year.

Month has been renamed Calendar Month.

A few supporting classes have been introduced for annotating
calendars: CalendarDate, AnniversaryDate, TimeAndDate.  The last of
which will probably find multiple other uses.  These are all
subclasses of AbstractTimeStamp.

Restrict number of days per year for planets to be >=
MIN_DAY_PER_YEAR.  Suggested by Gavin so that months continue to have
enough days to retain meaning as lengths of seasons change.

-----------------

NOTE:

This commit is 1 of 3 separate steps.  This first does not deactivate
the original call to ORBIT() and should preserve answers bitwise.
There is some risk thate deactivating that call to the original logic
will introduce roundoff in the calculation of zenith angle and is thus
being committed in the repository as a separate isolated step.  A 3rd step
will then delete the unused code.

Note - also included in this change, a correction in expression for
omega - from G. Russel.

Minor correction.