The MODELE.f coding to automatically adjust DT such that DTsrc/DT is an integer was removed, since it was already present in init_ATMDYN() (the latter also has the correct constraint that the ratio is an EVEN integer).
The automatic adjustment is enabled for planets other than Earth.
For Earth, the necessity to adjust will likely only occur if someone has
made a rundeck typo, and the user is notified via a call to stop_model
as before.

The rounding of DT_XUfilter and DT_XVfilter that was present in
MODELE.f is not strictly necessary.