This is basically a proof of concept. I have implemented a function which returns the list of files in a directory. Since there is no simple Fortran way to do this, this funtion relies on a call to the C system library.
One can use it as follows:

      use SystemTools
      character(len=80), dimension(16) :: flist
      integer :: i, num

      call stFileList("directory_name", flist, num)

      do i=1,num
        print *, trim(flist(i))
      enddo

Here "directory_name" is the name of the directory one wants to scan.
And the resulting list of files is returned in flist. num is the
number of values in flist. Result will include both, files and
directories, in particular, it includes "." and "..". The array
flist is expected to be allocated by user to a sufficient size
(if its size is too small the call will generate an error).

This code requires gfortran 4.8+ (not tested yet with intel,
hope it works).