Change to facilitate adding adding/subtracting layers that are not near the model top.
sync_param logic for VSDRAGL is now such that it is only necessary
to specify values near the model top.  Previously, all LM-LS1+1 elements
had to be specified, most of which are typically zero.

If VSDRAGL is set in the rundeck, but with a length less than
LM-LS1+1, the unspecified values are assumed to be zero.
The specified values fill the topmost layers in bottom-up order.