Various modifications to transmission correction factors; inclusion of water vapor absorption in the window region.
Changes to the dimensions of the correction-factor lookup tables:
(a) the number of levels was increased from 24 to 43.
(b) the number of column water vapor amounts was reduced from 15 to 9.
(c) the number of other absorber amounts was reduced from 11 to 7.
(d) the ratio of successive absorber amounts was increased from sqrt(2) to 2.

The choice of 43 levels is optimal for 40-layer configurations, but should
also yield improvements for higher resolutions.

To facilitate future changes to these dimensions, I introduced Fortran
parameters to replace the existing hard-coded 24's, 15's, and 11's.

Since water vapor profiles may have a vertical structure significantly
different than that of the reference atmosphere that was used to compute
the lookup tables, a per-layer and per-direction interpolation
within the lookup table is now performed.  Each interpolation
for downward fluxes at a given level selects the reference
atmosphere having the same column amount above that level
(and for upward fluxes, the column amount below that level.)
In cases for which water vapor mixing ratios increase upward
in the lower troposphere, an additional adjustment is performed.