sumfiles: file name length and exit status
sumfiles did not work with long file paths (e.g., when providing files with a long absolute path) because the file name was limited to 80 characters. Changed it to 4096, which is the maximum path name length allowed under Linux. Also provided exit status so calling program can know if something went wrong.