A few fixes to make the model run with TRAPS.
This is mostly related to configuration which reads
monthly MODIS LAI from a file. One may still need to
hack MODIS vegetation heights in the cells which
don't have vegetation to pass TRAPS.

canopyspitters.f a few variables were not defined if lai==0
(were used only in diags)

GHY.f - for lai==0, water holding capacity of canopy is zero,
which requires special treatment to avoid zero division

ENT_DRV.f, GHY_DRV.f - one of most stupid bugs of ifort 14. :
when compiled with traps was not treating the bounds of
FOCEAN correctly only because the same array was aslo passed
as an argument, even with a different name. weird...