This is a major commit, if you are running TCADI please read carefully!
Extracting seasalt aerosols out of the Koch aerosols. Now the Koch aerosols
have SO4,BC,OC only, and seasalt are controlled by the TRACERS_AEROSOLS_SEASALT
directive. There are also two files that need to be added in the rundecks:
TRACERS_AEROSOLS_SEASALT.F90 and SeasaltTracersMetadata.F90. The template
rundecks are all updated, but PLEASE REMEMBER that you will need to add
three lines in your preexisting rundecks to get all aerosols again. See
e.g. templates/E4TcadiF40.R and templates/tracer_aerosols_source_files
for the changes that need to be introduced.

==================================================
Remember, this change only applies to master branch, it is not intended to
be added in AR5_v2_branch unless you give me a very good reason to do so. The
results do not change (see next paragraph) but the changes are too numerous
to be ported in AR5_v2_branch for no reason.
==================================================

There are no changes in the results, other than roundoff errors. The exact
location where this roundoff occurrs is pinpointed; let me know if you care
about it and need to put a hack to exactly reproduce the original output.