Applied a cherry-pick related to snow albedo (form AR5_v2_branch). Still have to make it compatible with new atm-land interfaces.
Changed snow masking algorithm in ALBEDO module so that it uses
real depth of the snow (m) rather than snow water equivalent.
Also changed it to more explicitly use snow fraction information
passed by land surface. Both of these reduce the fraction of
exposed snow and help to melt it on the north of siberia by
the end of the summer.

Also added functionality to extract vegetation hights from
Ent structure for each pft. These are passed to 16-pft - 8-pft
conversion program and can be used for fine-tuning of
this conversion. For current setup this has no effect.

Conflicts:

	model/GHY_COM.f
	model/GHY_DRV.f
	model/RADIATION.f
	model/RAD_DRV.f