Added new options to QSUB_STRING functionality. Now one can use symbols %np and %t in the QSUB_STRING to pass to it the number of cores and execution time from the runE script. The corresponding values are passed to runE as command line arguments:
 -np <numper_of_mpi_trreads>
 -t  <execution_time>

So, if one has the following QSUB_STRING in ~/.modelErc :

  QSUB_STRING="sbatch --qos=debug -A s1001 -n %np -t %t -C sand"

then the command

  ../exec/runE Efoo -cold-restart -np 45 -t 0:15:00

will submit "Efoo" run to "debug" queue requesting at least 45 cores
(i.e 3 Sandy Bridge nodes) for 15 minutes. The default values
(if not specified on command line) are 1 core and 10 minutes.

If you are using setup-run target, then your NPES variable will
be passed as %np. There is no option for setup-run to pass %t .

If you want to use this functionality, then keep in mind that your
older versions of modelE will not understand these symbols, so
you will need to edit QSUB_STRING when you swithch between
new/old versions or leave old QSUB_STRING as it is and create
a new QSUB_STRING with some suffix and specify it on command line, i.e.

  QSUB_STRING_new="sbatch --qos=debug -A s1001 -n %np -t %t -C sand"

and use the command

  ../exec/runE Efoo -cold-restart -np 45 -t 0:15:00 -s new

Also, keep in mind that in these examples I use sbatch functionality
which allows me to just request a number of cores (with -n ...)
and let slurm decide how many nodes are needed. runE doesn't know
anything about computer architectures, so it will not calculate
the number of nodes explicitly.