Fix for variable orbit parameters.
Previous changes to treatment of calendar/orbit introduced an ordering
difficulty in which the orbit needs to know the current year before
"year" is defined by the calendar.  This does not arise with fixed
orbits, which do not depend on year.

This fix pushes the orbit%setYear() call to the end of initialization
when the year is available.