Eliminated reliance on F2003 automatic allocation of arrays on an assignment.  Although convenient, there are cases where misuse of this capability may result in bugs that are difficult to analyze.
Also eliminated the -assume realloc_lhs Intel compiler flag that was
needed to enable standard behavior in this regard.

There may be additional locations in the code that will turn up with
more testing, but the unit tests only required this one change.

This issue should be revisited when the F2008 standard becomes widely
portable, as automatic assignment of polymorphic entities may be desirable.