Build process change: building out of source. Since there was so much pressure recently, I have modified the build scripts to allow building of the model outside of source tree (i.e. putting .o and .mod to different directories, separate for each rundeck).
I did it in a rather rough way at the moment, so it is possible
that something will not work as expected. Use at your own risk
and let me know if you see problems. I'll clean it up later,
when I have time.

Those who don't want this functionality can stop reading now.
The default behaviour of the build system didn't change. But let
me know if you think that something got broken.

To enable this functionality specify BUILD_OUT_OF_SOURCE=YES
on a compile line (and with all other related "make" commands)
or add it to your ~/.modelErc .

Once it is enabled, "make setup ..." will create a directory
<run_id>_obj in decks directory and store all .o and .mod files
under this directory (separately for each rundeck). When doing
"make clean" make sure to specify rundeck name with RUN=... ,
so that make knows which directory to clean, or just delete
the corresponding *_obj directory.

If you want to work completely separately from the source directory
tree (for example, on a different file system), do the following.

1. create a directory where you want to work (this will be a remote
equivalent of "decks" directory).
2. symlink modelE/decks/Makefile to this directory
3. set an environment variable MODEL_E_ROOT to the location of
your modelE directory with the source code
4. work as described above (with BUILD_OUT_OF_SOURCE=YES set).
Build files (.o, .mod) will be stored under <run_id>_obj in
your working directory.

There is one exception to this "build out of source" rule:
Fortran files produced by the call to m4 are still stored inside
the source tree, since they have to be eventually committed to
the repository.

You may want to know, that makefiles are copied to build tree
only once (when one does the first setup). So if you modify
a makefile in the source tree, you may want to delete the build
directory to force it to get rebuild (simple "make clean" will
not do it).