Manual cherry-pick of AR5_v2_branch commit 584b98ffb525d829e6ba377581e41a38f05713eb (for VMP).  Automatic cherry-pick was not possible due to avant-necessaire conversion of clouds code to free-format.  Hopefully I have not made any merge mistakes.  No change to results unless use_vmp=1 in rundeck.  Have not yet tested climatology with use_vmp=1.
Aside from potential use in production runs, one additional motivation for porting VMP
to the master branch is to have a second "poor man" microphysics option against which
to compare the more sophisticated soon-to-be introduced scheme.  Some of the additional
arrays exported from the clouds to the radiation scheme as part of VMP may be renamed
or generalized once true mixed-phase physics arrives.

Log message from AR5_v2_branch commit:

Initial coding for virtual mixed-phase (VMP) clouds.  No effect on results unless opting in.

The VMP option is activated by setting use_vmp=1 in the rundeck.  VMP has been
introduced to the AR5_v2 branch on a trial basis and will be either
(1) improved as needed or (2) withdrawn if it creates more problems than it solves.

In short, VMP makes all-liquid clouds and all-ice precipitation for
Tmin_water<T<Tmax_ice, and the autoconversion rate of liquid condensate
to ice precipitation is interpolated between pure-ice and pure-water values.

Mixed-phase processes are "virtual" in the sense that the Model E
structure does not (yet) permit coexistence of liquid and ice budgets
in one gridbox.  VMP attempts to crudely express the following
two premises:

- Liquid droplets are "small" and ice crystals are "large".  Liquid lives
  longer and dominates radiatively.  Ice falls out quickly and has a smaller
  optical depth for a given water content than liquid.

- The importance of ice formation/sedimentation as a condensate sink
  can be expressed as a simple function of temperature.

For Tmin_water=-35C < T < -5C=Tmax_ice :

LHX=LHE, LHP=LHS.  For T<Tmin_water, LHX=LHS as before.

CM (the reciprocal of the local time constant for "autoconversion") varies with
temperature between the pure-ice value at T=Tmin_water and the pure-water value at
T=Tmax_ice.

per-file notes:

CLOUDS2.f
Aside from the new if(use_vmp) blocks, LSCOND has been slightly rearranged to try to
aggregate the "macro" and "micro" physics into contiguous sections.  The rearrangment was
not strictly necessary for VMP, but reduced the number of instances of if-tests on the
microphyics choice and should improve readability.

CLOUDS_COM.f
Now declares a second instance of TAUSS and CSIZSS to allow the radiation to see the ice that
coexists with liquid.

CLOUDS2_DRV.f
If use_vmp is .true. the cloud driver
(a) Fills the 3D arrays for the optical properties of the ice that coexists with water.
(b) Augments the ice water path diagnostic with WMPR over the model levels where LHP=LHS.

RAD_DRV.f
The core radiative transfer code was already built to handle coexisting liquid and ice, so it
was straightforward to place the second optical depth exported from the cloud scheme into TAUIC
(and SIZEIC).