Enabled assignment to dp reals from rationals.
real(kind=dp) :: x
type (Rational) :: r

x = r

This will enable the same for Time and TimeInterval, which should make
it easier to use the new time classes in the model.

This logic has been in the code for a while, but there is an odd Intel
bug that is exposed when doing the same for single precision.  By
commenting out the single precision interface (which we don't really
need), it seems to be fine.