Just for the purists who want to compile out of source "cmake way", I have added a script modelE/exec/configure.
To compile out of source, go to a directory where you want to
do it and execute

  path_to_source/modelE/exec/configure <RUN_NAME> <RUNSRC>

WHere  <RUN_NAME> is the name of the run you want to build
(rundeck name), <RUNSRC> -name of the template.

This will configure your build directory (will create a Makefile).
You can then build the model with simple

 make setup

If you want to pass extra options to "make setup", append them
to the configure command, i.e.

 path_to_source/modelE/exec/configure Efoo E4F40 MPI=YES COMPILE_WITH_TRAPS=YES

They will be remembered. To build the model you will still execute

 make setup

Run the model the usual way, using runE or Reto's script.