Duplicate tracer and find tracer by name
Added an optional argument to the creation of a new tracer with which an existing tracer can be specified to be used for initial values instead of the default initialization.

Added a function that returns the index of a given tracer.

(this commit completes the incomplete commit from 3/19/15 that did not update the m4 file)