Expect no change in results for normal rundeck conditions.
For anyone who wants to try out daily frequency NINT ozone input
I am pushing *experimental* code to try that out. One sets
define HIGH_FREQUENCY_O3_INPUT in the rundeck and lists either a
single input file or a directory of YYYY.nc files in the rundeck
as "OxHFfile". (E.g.standard timestream nc format daily frequency
file.) But this is a bit different input than the climatological
(monthly) ozone. This is on model levels 1,LM and remains in units
of volume mixing ratio. The file is expected to contain also the
surface pressure concurrent with the ozone supplied. Online in the
GCM the units will be converted from vmr to cm-atm and then using
REPART() readjusted a bit in the vertical to account for the current
model surface pressure vs. that in the input file. See comments in
the code. Note that the normal NINT climatology is still read --
this "high frequency" ozone just overwrites it on GCM levels. This
is so one does not have to decide on the radiation equilibrium
levels' ozone.

The reason I set it up this way is that the most common use case
would probably be outputting ozone from a TCADI run and inputting
that into a NINT run. To this end a script is provided:
aux/convertSUBDDtoNINT_O3.ksh
that would do the conversion to either of the 2 main timestream
input formats... The tracer run would have these settings:
  (define NEW_IO_SUBDD)
  define CACHED_SUBDD
  SUBDD= 'p_surf Ox'
  NSUBDD=48 (for 1/2 hour step of course)
  write_daily_files=1

Other notes:

Again, please test this out before using in a production run!

We could change in the future to still work if you don't have
surface pressure input.

Probably should some day relax the condition that input is on
same model levels (e.g. allow vertical regridding from one model
level configuration to another online).

The high frequency overwrite of climatology is done before *tracer*
values are allowed to do the same. So under a normal TCADI run
the high frequency ozone would be overwritten by tracer. But one
could imagine a setup where Lmax_rad_O3 is set to < LM, so you could
have climatological (monthly) ozone in the rad. equil. levels, tracer
ozone from 1,Lmax_rad_O3 and high frequency (daily) ozone from
Lmax_rad_O3+1,LM...

If someone wants to replace the CPP directive HIGH_FREQUENCY_O3_INPUT
with a db parameter to make the code cleaner, that's fine with me.
I didn't feel confident in setting it up this more modern way.

I am told that temporarily on the master branch the dir/YYYY.nc timestream
methodology need fixing, so only single-file method will work
with this new ozone input right now.