Resumed cleanup of OCN_Interp.f to reduce its required degree of specific knowledge of the inner workings of the rest of the model, e.g. the ocean equation of state.  This helps to streamline coupling and to reduce the amount of present and future code duplication.
This batch of updates changes results very slightly when
atm/ocean grids differ (which is why it is being done separately).

The GISS ocean model now computes various fields on its
own grid which in turn are regridded to the atmosphere by
OCN_Interp.f.  The primary contributor to differences in
results to the previous coding is SST which is calculated
pre-regridding rather than post-.
The nonlinearity of the equation of state is modest enough
that the new method changes the atm-grid SST by at most
~.001 K in testing 288x180 -> 144x90 regridding starting
from Levitus initial conditions.  A second contribution
to slight differences in results is that the first-layer
ocean mass has been eliminated as a regridding weight
(after all, the atmosphere wants to know surface fields).

At the level of roundoff differences, the time-mean ocean
surface height is computed pre-regridding rather than after.

Finally, the "bundled" version of AG2OG_oceans was made
consistent with an update to the unbundled version
concerning ice masking of the surface stress.