Additional comment lines are inserted into AtmL40.F90
AtmL40A.F90 is an alternative vertical layering with a mixture of constant and variable mass.
It was run in atmosphere LL mode for 12 years.
With no tuning, RMS1 comparisons for AtmL40A are worse than using standard AtmL40 .

AtmL100.F90 uses a mixture of constant and variable mass for vertical layering.
The model runs with an additional modification to the correction factors in RADIATION_ALT.f

SCM and tracers will soon be available for varaible mass, but not yet.