Allowing nitrate aerosol to have indirect effects.
This changes results of OMA runs.

PS. I do not understand why all solubility factors fq_are are hardcoded, and the fq_aer-variable is commented out.
See example below.. I didnt change that as I dont know if someone had a reason for doing so.
But I changed the hardcoded numbers for BCB to match what is used in the model.

      SSM5 = 1.70d12*DSU(5)*0.8!*fq_aer(n_OCB) ! OCB
      SSM6 = 1.70d12*DSU(6)         ! BCIA aged industrial BC
      SSM7 = 1.70d12*DSU(7)*0.8!*fq_aer(n_BCB) ! BCB
      SSM8 = 1.70d12*(DSU(18)*0.8!*fq_aer(n_isopp1a)
     &               +DSU(19)*0.8!*fq_aer(n_isopp2a)
     &               +DSU(20)*0.8!*fq_aer(n_apinp1a)
     &               +DSU(21)*0.8!*fq_aer(n_apinp2a)