Importing layering-related updates from the V2 branch. No effect on results.
Continuing to chip away at the set of issues causing code duplication
and other inconveniences when changing ocean layering:

(1) Added coding for the optoin to cold-start from an input netcdf file
containing T/S on the model horizontal grid but with arbitrary layering.
This streamlines changes to layering, bathymetry, etc. and makes
specification of idealized initial conditions more intuitive compared
to the default prodecure. The core of this coding was taken from the OIC
program in the init_cond directory and therefore the initial state for
Levitus-launched runs is identical to within roundoff except where missing
data is filled in a different way.

(2) Formerly hard-coded layer indices for the depth of solar penetration
and glacial melt deposition are now determined from physical depths
during initialization.

(3) Introduced a new method to select ocean layering via the rundeck
from a set of predefined choices; also split vertical and horizontal
layering into separate files.

(4) NOCEAN is now a rundeck parameter (just in case).