As discussed at GCM meetings and in various e-mails with folks, making the two changes below that change the answer in simulations with Shindell chemistry on. A third (smaller) change involving only NOx from lightning will follow soon, at which time I will inform Larissa and Jan for retuning (sorry).
Soon, I will propagate these same changes to the AR5_v2_branch for consistency.

(1) In template decks removing the AR5_FASTJ_XSECS CPP directive and changing
the SPECFJ file (chem_files/jv_spec_AV_X68d.dat) accordingly so that the newer
photolysis cross sections are used by default. The option remains to run the AR5
versions and the file to do so was left as a comment.

(2) passing chemical O3 instead of tracer Ox to the radiation code on levels
that perform chemistry. This is physically correct and improves some temperatures
high up the atmosphere e.g. in 96L and 102L models we tested. Including some new
diags for this so increasing KAIJ too. To go along with this, setting
templates/chemistry_params PltOx parameter to 0. hPa. This effectively turns off
the TOA {}Ox-species overwriting in the chemistry by climatology.