Added option to use the cyclic tridiagonal solver in ODIFF instead of the locally time-explicit treatment of east-west fluxes right at the dateline.
Note that this commit introduces roundoff differences for the case when the
new option is NOT used.  (To reduce the number of ifdefs, RU and RV are
built up in a different order at the dateline.)

The new option is not yet the default since it hasn't been tested much.
Presumably only the 4x5 ocean has an AKHMIN close enough to molasses that
it would be much affected by the new option, and then only for rapidly
time-varying conditions (which did generate a 4x5 crash recently).

The cyclic treatment is not (yet) being applied in the north-south direction
since it would likely have even less impact than in the east-west.