To facilitate ongoing development of the stratiform clouds routine, ( CLD_AER_CDNC || BLK_2MOM ) code blocks were moved from LSCOND to separate subroutines.  No differences to results for CADI, CAMPI prognostic variables.
This commit removes the never-recently-used ability to run BLK_2MOM
without CLD_AER_CDNC.

Unfortunately there are occasional differences in certain CLD_AER_CDNC
diagnostics as a result of moving their accumulations out of LSCOND into
the clouds driver, and attempting to capture the same information via
available saved arrays, but the overall patterns of output look very
similar to my eye.

While I did my best to include TOMAS in this code transfer, it was not
possible to properly test TOMAS since the baseline TOMAS is not currently
able to run with traps.  This commit caused a non-propagating TOMAS
difference after 1 hour in the ncl variable.

Thanks to Andy A and Susa B for assistance on various matters.