Imported from AR5_v2_branch for upcoming L102 runs on master until high-top radiation lookup tables are generated:
Added a CPP option to force the upward-flux transmission correction factors toward unity
near the model top.  This is necessary, for example, if the model top is above the pressure
domain for which the lookup-table factors were computed.  The range of pressures over which the
adjustment is applied is tunable, but the impact of the choice is modest compared to the
differences in behavior between adjusted and unadjusted factors.  Currently, the center
of the pressure range is set to 0.1 hPa.