Possible fix reproducibility with decaying tracers.
Noticed that having a tracer that undergoes radioactive decay
caused a reproducibility problem. I am not sure this fix is
the intention/is ideal, so someone who uses such tracers, please
check.

It looked to me expdec(n) array was being reset to 1 each call
and then only set to the per-tracer value if ifirst==.true..
It now calculates expdec each call.

expdec(n) array is declared as "save" on v2 branch,
but my compiler wouldn't let me do that when I tried here on master.
If that's preferable, perhaps someone knows how to do that instead.