Retune Shindell chemistry for F40 with VMP clouds.
Please take note:

Longer tests are ongoing, so it's not out of the question we'd
tweak this in next 1-2 weeks, but we don't anticipate that.

Changes include both deck settings and code changes, so you must
update in-sync for TCAD(I) runs!

Non-VMP settings are changed too, to new parameters, but are
just set equivalent to the old settings; these are commented out
in the decks.

(Tuning was done on master branch)

Also changing these parameters' defaults in the code to match those
committed here (F40 VMP) to reduce risk of mistakes.