Cleaning of the "make" scripts:
- more flexible implementation of "out of source" builds
- removed some dead/obsolete code
- added more flexibility to specification of source files
  in the components (now they can be put into subdirectories)

Also cleaned some cpp instructions related to include files.

Due to the intel compiler bug had to add couple of -I options
in model/shared/Makefile.

Inside the components (i.e. subdirectories under modelE/model)
one can now put the code into subdirectories. The source files
still have to be listed as before (without the path)
in FSRCS, F90SRCS, CSRCS. The directories where they are located
should be listed in SRC_DIRS (space separated), relative to
the component directory. The default is the component directory
(as before).
Also added an option to specify external files for the component
(since some seem to be doing it as a hack anyway). Just use
ABS_SRC_DIRS instead of SRC_DIRS and set it to the absolute path
to the external source files (space separated if more than one path).

Please keep in mind, that all object files in the component
will be stored in the same directory, so if code is located
in more than one subdirectory, there should be no source files with
identical names in these subdirectories.