Adding CO isotopes: C18O, C17O, 13CO. They are enabled by #define TRACERS_dCO
For now, this commit is a placeholder. No change in results, and the CO
isotopes just float around, there is no chemistry yet to form/destroy them.
Emissions are implemented, and work in the same way as in CO. For initial
conditions, for now, they use the same file as CO. Much more to come in the
coming months.