First step in expanding aircraft sources for CMIP6, as looks like we will have this source for more tracers now. Other commits on this to follow.
Cleanup of subroutine get_aircraft_tracer, in TRACERS.f, which now
gets passed the tracer index and the filename so that tracer-specific
coding (and looping over tracers) could be removed from inside the routine.
Incidentals:
  - for binary input file, rundeck parameters aircraft_Tyr1 and aircraft_Tyr2
    are now mandatory (safer, I think)
  - some rearrangement of the logic that picks the year based on
    aer_int_yr or o3_yr depending on the tracers compiled.
  - cosmetics

TRACERS_DRV.f: now more calls to subroutine get_aircraft_tracer as a
consequence of the changes to that routine. For clarity, grouping the
{zeroing of tr3Dsource(nAircraft), the call to get_aircraft_tracer,
and the apply_tracer_3Dsource(nAircraft) call together in blocks for
TRACERS_SPECIAL_Shindell, TRACERS_AMP, TRACERS_TOMAS, and
TRACERS_AEROSOLS_Koch.

Not expecting change to results in this commit. Will test.