Introduce do_aircraft(ntm) array of logicals, analogous to do_fire. Which tells the model when to expect/read/apply a 3D aircraft tracer source.
It is set to true when it find a file of the form trname()_AIC (for
backwards compatibility with the existing, hardcoded, sources).

In OldTracer_mod.F90 and OldTracer_mod.m4F90 I have simply replicated
the do_fire coding that Tom Clune et al. have done, though I don't
understand what it is all for.