Allow tracer emissions source attribute to be read from variables, not just global.
For netCDF emissions files the global "source" attribute was
read to set up diagnostic names for an emissions source. With
this change, the model *first* checks for source attribute on
the variable in question (e.g. "BCII"). If it doesn't exist,
the model will check for a global attribute (for backwards-
compatiblity with existing emissions files).

This small change allows us to group, e.g. each tracers'
emissions for a given sector in one file, taking advantage of
at least that part of timestream methodology.