(1) NTM_chem_{beg,end} initialized to 0 Please, if anyone (Kostas, Tom, Carlos?) disagrees that this should be the case, let me know.
(2) Allow abs(o3_yr) to control cyclic argument to init_stream call when appropriate.

In subroutine readSurfaceSource, it looked to me that the "cyclic" argument to the call
initializing tracer emissions timestreams was only dependent on aer_int_yr (or master_yr).
Really, in case one sets O3_yr and aer_int_yr differently, which parameter takes control
should depend on the tracer in question. This commit attempts to fix that. [For the moment,
until a tracer object is changed to "know" it is a chemistry tracer, this required passing
isChemTracer logical from TRACERS_DRV.f down to readSurfaceSources/readSurfaceSource
routines, based on NTM_chem_{beg,end} range criterion.]

So, *as I understand it*, you should get a change in results when/if:
(1) you list a netCDF emissions file in your deck AND
(2) it is multi-year nc file AND
(3) you set o3_yr to something different than aer_int_yr. This would include if you, e.g.,
    set aer_int_yr to something other than master_yr, but allowed o3_yr to default to
    master_yr.

I ran a few basic tests and that seemed to be the case.