Change template transient emissions files from those ending in 2000 to those ending in 2020. OMA, MATRIX and TOMAS. This seems preferable to me. I can change back if you disagree. Note: your simulation may change with this update if you are running model year 2000 or later, due to interpolation with later emissions. I.e. no change in results for templates set to 1850.
For now, when changing the aircraft emissions files (as above) also need to set:
aircraft_Tyr2=2020. So, I changed that in the templates.

Commenting out three input files that are only needed for rare
use case of coupled_chem .ne. 1. This is a step towards less tracer
deck changes when changing model year.

Removing a few variables from a use statement that are not used.

The changes to model/shared/RunTimeControls_mod.F90 are not mine so
I assume are the result of a pull? They have to do with Kostas' new
dCO tracers.