Please read carefully if you use aerosol tracers!
Implementing diag_fc, which replaces AMP_DIAG_FC and TOMAS_DIAG_FC, and
is also active for OMA. It can get two values:

diag_fc=1: radiation is called only once, collectively for all aerosol tracers
diag_fc=2: radiation is called several times (slow!), once per radiatively
           active aerosol tracer

Optical depth values are not affected; they are always calculated per
radiatively active tracer, this only affects swf and lwf diagnostics.

For TOMAS, the default is 2, since (based on the comments in the code) the
option of 1 is not implemented yet.

For OMA and MATRIX, the default is 1. CAREFUL, for OMA this is the
opposite from what it was in the past. This saves 16 radiation calls, but
only produces one bulk radiation diagnostic. Its value is not exactly
the sum of the individual ones, due to non-linearities. An one-day simulation
shows <1% difference for OMA and ~3% for MATRIX, but this is not accurate.
I will try to present how much of a difference it makes in the next modeling
meeting. The bulk value is now saved (randomly selected) to swf_SS (or lwf_SS),
only because this has the index 1. I will change that in the future, but it
requires a major restructuring in TRACERS_DRV.f.

The template rundecks of MATRIX and TOMAS produce the same results. OMA
rundecks are expected to have differences in the diagnostic variables swf_*
and lwf_* only.