Bugfix related with the aerosol indirect effect.
For OMA runs only, for cases that the Koch aerosols were not defined,
the aerosol indirect effect was not calculated, at least not as it should.
This promarily involves dust standalone runs, and maybe nothing else, but
can't be sure. I hadn't seen in that part of the code anything related
with minerals, so I believe that they are not implemented there.