Add code exploring a method of getting the same number of sunlight chemistry calls per day at one latitude. No change in results without new CPP directive engaged. The other day it was discovered that there are slight differences in longitude as to how many sunlight vs. darkness cases happen in the chemistry in one day. Seems to average out smoothly for monthly diags, so a significant change in climatology isn't expected, but it would be good to improve this for looking at higher time- frequency variables (e.g. SUBDD).
The commented-out:
"#define SMOOTH_SUNLIGHT_CHEMISTRY"
was added to the 3 maintained chemistry-related rundecks, along with a file ALB_IC
which is only read if this is turned on.

Should be no change in results without opt-in.

I see two problems:

(1) we were using ALB(I,J,1) from the radiation code as a criterion for sunlight
(along with COSZ1). Since ALB is only updated on NRAD steps, this was causing some
longitudes to have 1 less sunlit timestep, chemistry-wise, I think when the last-
remembered ALB() was 0 but this was too "out of date". This SMOOTH_SUNLIGHT_CHEMISTRY
workaround does not use the ALB(i,j,1) as a sunlight criterion anymore. The problem
with that is that fastj2 uses the albedo [I will try to find out from Drew if that
was the original reason for this ALB-condition]. So when ALB(i,j,1)=0., it is not
passed to fastj2. Instead we pass the last value ALB(i,j,1) had that was non-zero.
Not ideal, because this has to be saved in the rsf file (in mostRecentNonZeroAlbedo(i,j))
for reproducibility. I couldn't think of how to do this a better way without knowing
the ALB() from an RCOMPX call in the current timestep. An annoyance is that
mostRecentNonZeroAlbedo() (initialized to zero) will get filled in over most of the
globe within a half-day running, but in seasonal perpetual darkness areas, it won't
for a while. So, at TIME=ITIME1 *only* I am currently reading in the file 'ALB_IC',
which contains an annual mean value at each (I,J) so that something reasonable is used
until mostRecentNonZeroAlbedo() is completely refreshed by ALB(i,j,1)

Consequently, fastj2 now gets passed in surfaceAlbedo rather than using ALB(i,j,1)
from the radiation within. Without above directive on, ALB(i,j,1) is passed, to
replicate original behavior. With it defined, ALB(i,j,1) is passed if non-zero,
otherwise mostRecentNonZeroAlbedo(i,j) is passed.

(2) Even assuming (1) works, I think we will still get different number of sunlight
chemistry calls due to the SZA < some maximum criterion. For example, in the 144x90
lat-lon model, with DTsrc=1800, a half-hour is 3 longitudes. So the COSZ1 (physics
step-averaged cosine of sza) "pattern" is repeated one timestep later at 3 longitudes
over. But the two longitudes "in-between" will not have the exact same COSZ1 pattern.
Thus, with any fixed szamax cutoff of light/dark chemistry, it's a bit unpredictable
whether a time step near dusk/dawn will be counted as light or dark chemistry. (To
complicate things, we need to include twilight chemistry when 90 < sza < 98. but COSZ1
won't give us values > 90.) Anyway, this second problem has not been addressed yet,
as I don't know how.